refactor(#726): rewire imports, docstrings, and docs to the strategies-package homes
Test rewires: tests/unit/test_search_album_fuzzy.py and tests/integration/test_search_album_fuzzy.py import search_album_fuzzy from lookup.strategies.track_release_matching (the unit file's module docstring path updated too; both keep their lookup.orchestrator.lookup_releases_by_track patches because those intercept resolve_albums_for_track / search_compilations_for_track, which stay in the orchestrator); tests/integration/test_song_as_track.py and tests/unit/test_orchestrator_helpers.py import search_song_as_track from lookup.strategies.song_as_track; test_orchestrator_helpers.py and tests/unit/test_orchestrator_gaps.py import search_with_alternative_interpretation from lookup.strategies.swapped_interpretation and (gaps) search_album_fuzzy from the kernel module. No monkeypatch target needed retargeting: every remaining lookup.orchestrator.* patch site exercises code that stays (resolve_albums_for_track, search_compilations_for_track, search_song_as_artist, filter_results_by_track_validation, enrichment attrs), and the moved kernel reaches Discogs only through the injected service object, not module globals. Doc updates: docs/architecture.md strategy-pipeline intro, kernel note, and key-files list now name lookup/strategies/track_release_matching.py; docs/env-vars.md's LML_RESOLVE_NONLIBRARY_RELEASE entry places the SONG_AS_TRACK/SWAPPED kernel in its new home while TRACK_ON_COMPILATION stays in lookup/orchestrator.py until PR 4b; library/db.py's exact_title docstring and the strategies package/shell docstrings drop their stale orchestrator paths.
