+++ docs/architecture.md
@@ -18,7 +18,7 @@
 
 ## Search Strategy Pipeline
 
-The strategy seam (`Outcome` value type, `Strategy` Protocol, `_apply` write site, `execute_search_pipeline` runner) lives in `core/search.py`. Each concrete strategy is a `@dataclass(frozen=True)` class in `lookup/strategies/`, one module per strategy; the orchestrator-side execute funcs that do the actual library/Discogs work still live in `lookup/orchestrator.py` and are injected into each strategy at construction time.
+The strategy seam (`Outcome` value type, `Strategy` Protocol, `_apply` write site, `execute_search_pipeline` runner) lives in `core/search.py`. Each concrete strategy is a `@dataclass(frozen=True)` class in `lookup/strategies/`, one module per strategy; the execute funcs that do the actual library/Discogs work are injected into each strategy at construction time. `SONG_AS_TRACK`'s and `SWAPPED_INTERPRETATION`'s execute funcs live in their strategy modules (over the shared kernel in `lookup/strategies/track_release_matching.py`); the other three still live in `lookup/orchestrator.py` (decomposition in progress, LML#722).
 
 | Strategy | Trigger | Class | Execute func |
 |---|---|---|---|
@@ -30,11 +30,12 @@ The strategy seam (`Outcome` value type, `Strategy` Protocol, `_apply` write sit
 
 Strategies never mutate `SearchState`; they return an `Outcome` and the runner applies it via `_apply`. This makes cancellation safety structural — a cancelled `attempt()` is a no-op against state by construction. Production wiring lives in `lookup.strategies.build_strategies`, called from `perform_lookup`.
 
-`SWAPPED_INTERPRETATION` is **Discogs-touching** (LML#622): once it identifies the artist side, it cross-references the *other* part as a track and narrows to the release(s) that actually contain it, rather than returning the artist's whole discography. The release→library matching reuses the same kernel as `SONG_AS_TRACK` (`_match_track_releases_to_library` — Discogs track search → `search_album_fuzzy` → deferred tracklist validation, bounded by `_chunked_gather` + a `MAX_SEARCH_RESULTS` early-exit). SWAPPED passes both `artist=` (filters the Discogs search) and `require_artist=` (re-filters the matched library rows to the identified artist), so the narrowing stays the artist's own release(s) and the keyword-supplement fallback can't leak another artist's release in; VA compilations are SONG_AS_TRACK's domain (it runs with `artist=None`). When narrowing fires it emits a `matched_via` map (an `Outcome.track_match`, like `SONG_AS_TRACK`); when nothing cross-references — or no `discogs_service` is wired — it returns the artist-filtered result via `Outcome.found` as before. Of the strategies, only `ARTIST_PLUS_ALBUM` is library-only.
+`SWAPPED_INTERPRETATION` is **Discogs-touching** (LML#622): once it identifies the artist side, it cross-references the *other* part as a track and narrows to the release(s) that actually contain it, rather than returning the artist's whole discography. The release→library matching reuses the same kernel as `SONG_AS_TRACK` (`_match_track_releases_to_library` in `lookup/strategies/track_release_matching.py` — Discogs track search → `search_album_fuzzy` → deferred tracklist validation, bounded by `_chunked_gather` + a `MAX_SEARCH_RESULTS` early-exit). SWAPPED passes both `artist=` (filters the Discogs search) and `require_artist=` (re-filters the matched library rows to the identified artist), so the narrowing stays the artist's own release(s) and the keyword-supplement fallback can't leak another artist's release in; VA compilations are SONG_AS_TRACK's domain (it runs with `artist=None`). When narrowing fires it emits a `matched_via` map (an `Outcome.track_match`, like `SONG_AS_TRACK`); when nothing cross-references — or no `discogs_service` is wired — it returns the artist-filtered result via `Outcome.found` as before. Of the strategies, only `ARTIST_PLUS_ALBUM` is library-only.
 
 ## Key Files
 
-- `lookup/orchestrator.py` -- Pipeline spine: `perform_lookup()` and the search-strategy implementations (decomposition in progress, LML#722)
+- `lookup/orchestrator.py` -- Pipeline spine: `perform_lookup()` and the not-yet-extracted strategy implementations (decomposition in progress, LML#722)
+- `lookup/strategies/` -- Strategy classes (gating predicate + `attempt` per module), the `build_strategies` factory, and the extracted execute funcs (`song_as_track`, `swapped_interpretation`) over the shared `track_release_matching` kernel
 - `lookup/matching.py` -- Pure matching predicates, filters, and score floors
 - `lookup/concurrency.py` -- `_chunked_gather` + the per-invocation Discogs API-call cap
 - `lookup/artist_resolution.py` -- Canonical-artist resolver (`resolve_canonical_artist`) + its flag gates and telemetry projections
